Parallax & Sorting Particles
Added parallax effect to the background attached to the camera.
Added a particle system to the player and a sorting layer exclusively
for particles. A script is needed for setting the sorting layer of a
particle system since Unity doesn't allow that.
At this point Unity changed some lines of code since the demonstration
in the videos used an old API